// 5-TrayElements/5-trayElements.d.ts
export declare const proyectElements: Readonly<{
    proyectAlmacen: HTMLUListElement | null;
    createTaskBtn: HTMLButtonElement | null;
    titleContainer: HTMLDivElement | null;
    objContainer: HTMLDivElement | null;
    empty: HTMLElement | null;
}>;
export declare const tasksElements: Readonly<{
    taskAlmacen: HTMLUListElement | null;
}>;
export declare const layoutElements: Readonly<{
    logo: HTMLDivElement | null;
    menu: HTMLElement | null;
    menuContainer: HTMLDivElement | null;
}>;
//# sourceMappingURL=5-trayElements.d.ts.map